Change how mouse position is stored - Store mouse pos as double - Reference MouseState values directly when getting cursor pos

+++ input/InputState.h
@@ -23,8 +23,8 @@ struct InputState {
   Keys lastKeys;
 
   struct Mouse {
-    float x_pos{0};
-    float y_pos{0};
+    double x_pos{0};
+    double y_pos{0};
   };
   Mouse mouse;
   Mouse lastMouse;
